errores corregidos de ts
